# SI-9776 Fix type of PriorityQueue.newBuilder and improve performance