review by @Ichoran

Mainly fixes issues from SI-9776:
- Fixes the return type of mutable.PriorityQueue.newBuilder to be Builder[A, PriorityQueue[A]] rather than PriorityQueue[A].
- Improves performance of bulk inserts from O(N log N) to O(N), using the heapify algorithm instead of one-by-one inserts.  This primarily affects the builder, ++=, and reverse.  (These changes indirectly benefit the many, many other methods that use the builder or ++=.)

Also fixes issue from SI-9757:
- Fixes space leak in dequeue.  (However, there is still a similar space leak in clear.)

Non SI improvements:
- Slightly improves performance of clone.
- Improves scaladoc description of reverse, which was already out of date from an earlier commit.

Informal timings of the speedups using scalameter:
- Creating a priority queue of 1 million increasing numbers: speedup from 3098ms to 151ms!
- Creating a priority queue of 1 million random numbers: speedup from 338ms to 133ms.
- Reversing a priority queue of 1 million random numbers: speedup from 590ms to 219ms.