@Ichoran 2016-05-21
Looks good; just a few minor comments.  Thanks!  This should speed PriorityQueue up a lot.