@chrisokasaki 2016-05-22
I made all of those changes (except the reusable builder).  The speedup to clone was huge, nearly a factor of 10!  I was also able to use some of those same ideas in reverse, but there the improvement is much more modest because it still needs to heapify.