@Ichoran 2016-05-22
Glad that made a big difference!  There's still the `pq != c` vs. `pq ne c` issue in the test, though (my last comment).  You want reference equality (the latter), right?