@chrisokasaki 2016-05-22
Sigh.  Yes, I made the change but then accidentally left the test file out of the commit.