@Ichoran 2016-05-22
I think ScalaCheck isn't running the test because it wants `object Test` (not `PQTest`).  Also, if you could squash all commits (no value in keeping them separate; the history is cleaner with a single commit--though of course you can make your own branch that preserves them), the PR should be set to merge.