@chrisokasaki 2016-05-22
Thanks, I couldn't make sense of the error reports.  I'll change the name and squash the commits.