@chrisokasaki 2016-05-22
FYI, changing the name (PQTest to Test) and squashing the commits in one go is giving me conflicts, so I'm going try doing that in separate steps, which means that there's going to be at least one meaningless intermediate update before I can get it right.  Just ignore it. :-)