@chrisokasaki 2016-05-23
Ok, here's the deal.  This is the first time I've ever used git (I normally use hg).  And like every git newbie, not only have I dorked things up far beyond my ability to fix them, but in my ignorant flailing around trying to fix things, I've also put things into an inconsistent state where these 6 commits CAN'T be squashed (or at least not by me).

The good news is that it's only 2 files.

From my end, the easiest thing to do would be to kill this pull request, and make a new one from scratch.  The main drawback is that it would lose the record of the conversation above.

Unless you have a better suggestion, that's my leaning.