@retronym 2016-05-23
@chrisokasaki Here's the recipe to update this pull request with a single commit: https://github.com/scala/scala/pull/5184#issuecomment-220868526