@chrisokasaki 2016-05-23
Awesome, that worked!  Thanks!